CID#982700 silence coverity

+++ sal/qa/osl/file/osl_File.cxx
@@ -6107,8 +6107,12 @@ namespace osl_Directory
         Directory::remove(buffer);
 
         i = rtl_ustr_lastIndexOfChar(buffer, '/');
-        buffer[i] = 0;
-        Directory::remove(buffer);
+        assert(i != -1);
+        if (i != -1)
+        {
+            buffer[i] = 0;
+            Directory::remove(buffer);
+        }
     }
 
     //########################################
